# Support `ExchangeType` on the client side