--- a/core/src/main/java/com/linecorp/armeria/client/HttpRequestSubscriber.java
+++ b/core/src/main/java/com/linecorp/armeria/client/HttpRequestSubscriber.java
@@ -103,7 +103,7 @@ final class HttpRequestSubscriber extends AbstractHttpRequestHandler implements 
     @Override
     public void onError(Throwable cause) {
         isSubscriptionCompleted = true;
-        fail(cause);
+        failRequest(cause);
     }
 
     @Override