… benchmarks/jmh/src/jmh/java/com/linecorp/armeria/client/WebClientIntegrationBenchmark.java
… core/src/main/java/com/linecorp/armeria/client/ClientRequestContextWrapper.java
… core/src/main/java/com/linecorp/armeria/client/DefaultBlockingWebClient.java
… core/src/main/java/com/linecorp/armeria/client/DefaultClientRequestContext.java
… core/src/main/java/com/linecorp/armeria/client/DefaultRequestOptions.java
… core/src/main/java/com/linecorp/armeria/client/DefaultWebClient.java
… core/src/main/java/com/linecorp/armeria/client/FutureTransformingRequestPreparation.java
… core/src/main/java/com/linecorp/armeria/client/HttpClientDelegate.java
… core/src/main/java/com/linecorp/armeria/client/HttpSessionHandler.java
… core/src/main/java/com/linecorp/armeria/client/RedirectingClient.java
… core/src/main/java/com/linecorp/armeria/client/RequestOptions.java
… core/src/main/java/com/linecorp/armeria/client/RequestOptionsBuilder.java
… core/src/main/java/com/linecorp/armeria/client/RequestOptionsSetters.java
… core/src/main/java/com/linecorp/armeria/client/TransformingRequestPreparation.java
… core/src/main/java/com/linecorp/armeria/client/UserClient.java
… core/src/main/java/com/linecorp/armeria/client/WebClient.java
… core/src/main/java/com/linecorp/armeria/client/WebClientRequestPreparation.java
… core/src/main/java/com/linecorp/armeria/client/retry/AggregatedHttpRequestDuplicator.java
… core/src/main/java/com/linecorp/armeria/client/retry/RetryingClient.java
… core/src/main/java/com/linecorp/armeria/common/AbstractHttpMessageBuilder.java
… core/src/main/java/com/linecorp/armeria/common/ExchangeType.java
… core/src/main/java/com/linecorp/armeria/common/FixedHttpRequest.java
… core/src/main/java/com/linecorp/armeria/common/NonWrappingRequestContext.java
… core/src/main/java/com/linecorp/armeria/common/RequestContext.java
… core/src/main/java/com/linecorp/armeria/common/RequestContextWrapper.java
… core/src/main/java/com/linecorp/armeria/internal/client/AggregatedHttpRequestDuplicator.java
… core/src/main/java/com/linecorp/armeria/server/AggregatingDecodedHttpRequest.java
… core/src/main/java/com/linecorp/armeria/server/DecodedHttpRequest.java
… core/src/main/java/com/linecorp/armeria/server/DefaultServiceRequestContext.java
… core/src/main/java/com/linecorp/armeria/server/EmptyContentDecodedHttpRequest.java