Motivation:

An optimized runtime implementation for non-streaming by sequence
writing `HttpObject`s followed by a batch flush.

Modifications:

- Add `AbstractHttpRequestHandler` for a common abstract layer between
  `HttpRequestSubscriber` and `AggregatedHttpRequestHandler`
  - `AggregatedHttpRequestHandler` is in control of the write of `AggregatedHttpRequest`
     if request streaming is disabled.
- Set a proper `ExchangeType` when sending an `HttpRequest` with
  a `WebClient`, a gRPC client or a Thrift client.
  - Thrift clients and `BlockingWebClient`s always use `ExchangeType.UNARY`.
  - Response streaming is automatically disabled if `ResponseAs.requiresAggregation()`
    is false.
- Add `ClientRequestContext.exchangeType()` to get the `ExchangeType` of the current
  request.
- Aggregate an `HttpRequest` in `RetryingClient` and `RedirectingClient` 
  if request streaming is disabled.
  - Add `AggregatedHttpRequestDuplicator` to provide an optimized
    dupliator implementation for unary requests.


Result:

You can now give a performance optimization hint to a client by setting
 an `ExchangeType` to a `RequestOptions`. An `ExchangeType` is 
automatically detected when a gRPC client, a Thrift client, or 
`WebClient` with `ResponseAs` is used.

Benchmark results
* Current branch
  ```
  Benchmark                                  (useRetry)   Mode  Cnt      Score     Error  Units
  WebClientIntegrationBenchmark.getRequest        false  thrpt   10  63934.425 ± 296.847  ops/s
  WebClientIntegrationBenchmark.getRequest         true  thrpt   10  46087.249 ± 190.453  ops/s
  WebClientIntegrationBenchmark.postRequest       false  thrpt   10  50190.294 ± 119.205  ops/s
  WebClientIntegrationBenchmark.postRequest        true  thrpt   10  31890.093 ± 153.674  ops/s
  ```
* Master branch
  ```
  Benchmark                                  (useRetry)   Mode  Cnt      Score      Error  Units
  WebClientIntegrationBenchmark.getRequest        false  thrpt   10  65629.231 ±  462.563  ops/s
  WebClientIntegrationBenchmark.getRequest         true  thrpt   10  54621.237 ±  630.048  ops/s
  WebClientIntegrationBenchmark.postRequest       false  thrpt   10  41906.188 ±  121.708  ops/s
  WebClientIntegrationBenchmark.postRequest        true  thrpt   10  33552.247 ± 1898.497  ops/s
  ```

Notable changes
- Improved performance on POST requests
  - The reduced invocations of `channel.flush()` from to 2 to 1 may affect the
    performance.
- Regression in `RetryingClient`
  - `HttpRequest.aggregateWithPooledObject()` is called twice after refactoring.
    This overhead may be eliminated by addressing
    https://github.com/line/armeria/blob/14e91d61c9750797feb7befe3f77cd9cfeb8a36e/core/src/main/java/com/linecorp/armeria/common/FixedHttpRequest.java#L47