@ikhoon 2022-06-14
### Benchmark results
* Current branch
  ```
  Benchmark                                  (useRetry)   Mode  Cnt      Score     Error  Units
  WebClientIntegrationBenchmark.getRequest        false  thrpt   10  63934.425 ± 296.847  ops/s
  WebClientIntegrationBenchmark.getRequest         true  thrpt   10  46087.249 ± 190.453  ops/s
  WebClientIntegrationBenchmark.postRequest       false  thrpt   10  50190.294 ± 119.205  ops/s
  WebClientIntegrationBenchmark.postRequest        true  thrpt   10  31890.093 ± 153.674  ops/s
  ```
* Master branch
  ```
  Benchmark                                  (useRetry)   Mode  Cnt      Score      Error  Units
  WebClientIntegrationBenchmark.getRequest        false  thrpt   10  65629.231 ±  462.563  ops/s
  WebClientIntegrationBenchmark.getRequest         true  thrpt   10  54621.237 ±  630.048  ops/s
  WebClientIntegrationBenchmark.postRequest       false  thrpt   10  41906.188 ±  121.708  ops/s
  WebClientIntegrationBenchmark.postRequest        true  thrpt   10  33552.247 ± 1898.497  ops/s
  ```

### Notiable changes
- Improved performance on POST requests
  - The reduced invocations of `channel.flush()` from to 2 to 1 may affect the
    performance.
- Regression in `RetryingClient`
  - `HttpRequest.aggregateWithPooledObject()` is called twice after refactoring.
    This overhead may be elimanated by addressing
    https://github.com/line/armeria/blob/14e91d61c9750797feb7befe3f77cd9cfeb8a36e/core/src/main/java/com/linecorp/armeria/common/FixedHttpRequest.java#L47